minor refactor to lib/src/features structure
* create run_commons/ directory for things shared between drive/ and test/

* add support for --device argument on Android
