+++ packages/patrol_cli/lib/src/features/drive/platform/android_driver.dart
@@ -3,8 +3,8 @@ import 'package:dispose_scope/dispose_scope.dart';
 import 'package:patrol_cli/src/common/artifacts_repository.dart';
 import 'package:patrol_cli/src/common/common.dart';
 import 'package:patrol_cli/src/common/logger.dart';
-import 'package:patrol_cli/src/features/drive/constants.dart';
-import 'package:patrol_cli/src/features/drive/device.dart';
+import 'package:patrol_cli/src/features/run_commons/constants.dart';
+import 'package:patrol_cli/src/features/run_commons/device.dart';
 
 class AndroidDriver {
   AndroidDriver({
