+++ packages/patrol_cli/lib/src/features/drive/platform/ios_driver.dart
@@ -5,8 +5,8 @@ import 'package:path/path.dart' show basename;
 import 'package:patrol_cli/src/common/artifacts_repository.dart';
 import 'package:patrol_cli/src/common/common.dart';
 import 'package:patrol_cli/src/common/logger.dart';
-import 'package:patrol_cli/src/features/drive/constants.dart';
-import 'package:patrol_cli/src/features/drive/device.dart';
+import 'package:patrol_cli/src/features/run_commons/constants.dart';
+import 'package:patrol_cli/src/features/run_commons/device.dart';
 import 'package:platform/platform.dart';
 import 'package:process/process.dart';
