+++ packages/patrol_cli/lib/src/features/run_commons/device.freezed.dart
