+++ packages/patrol_cli/lib/src/features/test/test_runner.dart
@@ -1,8 +1,7 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:patrol_cli/src/features/run_commons/device.dart';
 
 part 'test_runner.freezed.dart';
-
-enum Platform { android, iphoneos, iphonesimulator }
 
 @freezed
 class AppOptions with _$AppOptions {
@@ -10,12 +9,11 @@ class AppOptions with _$AppOptions {
     required String target,
     required String? flavor,
     required Map<String, String> dartDefines,
-    required Platform platform,
   }) = _AppOptions;
 }
 
 class TestRunner {
-  Future<void> run(AppOptions options) async {
+  Future<void> run(AppOptions options, Device device) async {
     throw UnimplementedError('`run()` must be implemented');
   }
 }
