+++ packages/patrol_cli/test/fakes.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 
 import 'package:mocktail/mocktail.dart';
-import 'package:patrol_cli/src/features/drive/device.dart';
+import 'package:patrol_cli/src/features/run_commons/device.dart';
 
 class FakeDevice extends Fake implements Device {}
